Change message timestamp from seconds to ms. Fill test chats from movie scripts

+++ src/dotnet/Messenger.Common/Elastic/Models/EsMessage.cs
@@ -9,6 +9,6 @@ namespace Messenger.Common.Elastic.Models
         public int ChatId { get; set; }
         public string Message { get; set; }
         public int UserId { get; set; }
-        public int MessageTime { get; set; }
+        public long MessageTime { get; set; }
     }
 }
